wip: ported reverb

+++ Inc/audio/reverbUtils.h
@@ -4,24 +4,24 @@
 
 typedef struct 
 {
-    int16_t coefficient;
+    float coefficient;
     uint16_t delayPtr;
     uint16_t delayInSamples;
-    int16_t oldValues;
+    float oldValues;
     uint16_t bufferSize; 
-    int16_t * delayLineIn;
-    int16_t * delayLineOut;
+    float * delayLineIn;
+    float * delayLineOut;
 } AllpassType;
 
 typedef struct 
 {
-    int16_t * delayPointers[4];
+    float * delayPointers[4];
     int16_t delayTimes[4];
     int16_t delayPointer;
     uint16_t diffusorSize;
 } HadamardDiffuserType;
 
-int16_t  allpassProcessSample(int16_t sampleIn,AllpassType*allpass,volatile uint32_t*);
-void hadamardDiffuserProcessArray(int32_t * channels,HadamardDiffuserType*data,volatile uint32_t * audioStatePtr);
+float  allpassProcessSample(float sampleIn,AllpassType*allpass,volatile uint32_t*);
+void hadamardDiffuserProcessArray(float * channels,HadamardDiffuserType*data,volatile uint32_t * audioStatePtr);
 
 #endif
